+++ packages/webapp/src/views/combined-case/check-and-submit.html
@@ -251,7 +251,7 @@
           { text: "Habitat" },
           { text: "Condition" },
           { text: "Size" },
-          { text: "Units" }
+          { html: "Biodiversity&nbsp;units" }
         ],
         rows: matchedHabitats
       }) }}
